requirements.txt
opencv-python>=4.8.0

ultralytics>=8.0.196

torch>=2.1.0; python_version<"3.12"

torchvision>=0.16.0; python_version<"3.12"

firebase-admin>=6.8.0

easyocr>=1.7.1

numpy>=1.23.0

PyQt6>=6.5.0
paho-mqtt>=2.0.2
requests>=2.31.0